Implementado o cliente REST, trabalhando com WebService SOAP e implementado cliente SOAP.

+++ controllers/pagamentos.js
@@ -85,30 +85,38 @@ module.exports = function(app){
 
                     var clienteCartoes = new app.servicos.clienteCartoes();
                     clienteCartoes.autoriza(cartao, function(exception, request, response, retorno){
-                        console.log(201).json(retorno);
+                        if(exception){
+                            console.log(exception)
+                            res.status(400).send(exception)
+                            return
+                        }
+
+                        res.location('/pagamentos/pagamento/' + pagamento.id)
+
+                        var response= {
+                            dados_do_pagamento: pagamento,
+                            cartao: retorno,
+                            links:[
+                                {
+                                    href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                    rel:"confirmar",
+                                    method:"PUT"
+                                },
+                                {
+                                    href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                    rel:"cancelar",
+                                    method:"DELETE"
+                                },
+                            ]
+                        }
+                        console.log(201).json(response);
                         return;
                     });
 
                     res.status(201).json(cartao);
                     return;
                 } else {
-                    res.location('/pagamentos/pagamento/' + pagamento.id)
 
-                    var response= {
-                        dados_do_pagamento: pagamento,
-                        links:[
-                            {
-                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                                rel:"confirmar",
-                                method:"PUT"
-                            },
-                            {
-                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                                rel:"cancelar",
-                                method:"DELETE"
-                            },
-                        ]
-                    }
 
                     res.status(201).json(response)
                 }
